* Bundles/ASFTags/ASFTags.m           some format string fixes         * CynthiuneSliderCell.m           float -> CGFloat         * Bundles/MP3/MP3.m           memcpy -> memmove, prevents segfault on playing MP3           on OpenBSD
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@3064 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Cynthiune/CynthiuneSliderCell.m
@@ -98,7 +98,7 @@
   return NO;
 }
 
-- (float) knobThickness
+- (CGFloat) knobThickness
 {
   NSSize size;
